Read command outputs and exit status through temp files.
Use separate files for command outputs and a common one for
command status. This works because commands in each suite have to
be ran sequentially.

This makes reading outputs back more straight forward at the expense
of more file I/O. This also requires the shell program to be able to
configure stream redirection in a separate command.