Leave context around each diff change.
Create a new Hunk type and construct a sequence of hunks from an input
sequence of changes accounting for overlap in change contexts.

This can likely be made faster at the expense of memory consumption if
we store the original test spec lines so that we don't need to
reconstruct them for .err file & diff output.